create local runner
