+++ sematic/BUILD
@@ -29,14 +29,19 @@ sematic_py_lib(
         ":function",
         ":resolver",
         ":retry_settings",
+        ":runner",
         "//sematic:future_context",
         "//sematic/future_operators:init",
         "//sematic/plugins/external_resource:timed_message",
         "//sematic/plugins/kuberay_wrapper:standard",
         "//sematic/resolvers:cloud_resolver",
         "//sematic/resolvers:local_resolver",
+        "//sematic/resolvers:silent_resolver",
         "//sematic/resolvers:resource_requirements",
         "//sematic/resolvers:worker",
+        # "//sematic/runners:cloud_runner",
+        "//sematic/runners:local_runner",
+        "//sematic/runners:silent_runner",
         "//sematic/types:casting",
         "//sematic/types:init",
         "//sematic/utils:exceptions",
