+++ sematic/caching/__init__.py
@@ -2,6 +2,6 @@
 from sematic.caching.caching import (  # noqa: F401
     CacheNamespace,
     CacheNamespaceCallable,
+    determine_cache_namespace,
     get_future_cache_key,
-    resolve_cache_namespace,
 )
