+++ sematic/caching/tests/BUILD
@@ -6,7 +6,7 @@ pytest_test(
         "//sematic:abstract_future",
         "//sematic:init",
         "//sematic/caching:caching",
-        "//sematic/resolvers:state_machine_resolver",
+        "//sematic/runners:state_machine_runner",
         "//sematic/types:serialization",
     ],
 )
