+++ sematic/caching/tests/test_caching.py
@@ -8,8 +8,8 @@ import pytest
 # Sematic
 from sematic import func
 from sematic.abstract_future import AbstractFuture
-from sematic.caching.caching import get_future_cache_key, resolve_cache_namespace
-from sematic.resolvers.state_machine_runner import StateMachineRunner
+from sematic.caching.caching import determine_cache_namespace, get_future_cache_key
+from sematic.runners.state_machine_runner import StateMachineRunner
 
 # these values were calculated by hand on paper to validate the algorithm
 # (they were all correct on the first try)
@@ -70,20 +70,20 @@ def test_namespace_str_happy_path(my_future: AbstractFuture):
 
 
 def test_resolve_namespace_str_happy_path(my_future: AbstractFuture):
-    actual = resolve_cache_namespace("my_namespace", my_future)
+    actual = determine_cache_namespace("my_namespace", my_future)
     assert actual == "my_namespace"
 
 
 def test_resolve_namespace_callable_happy_path(my_future: AbstractFuture):
-    actual = resolve_cache_namespace(my_namespace, my_future)
+    actual = determine_cache_namespace(my_namespace, my_future)
     assert actual == "my_namespace"
 
-    actual = resolve_cache_namespace(my_other_namespace, my_future)
+    actual = determine_cache_namespace(my_other_namespace, my_future)
     assert actual == "my_other_namespace"
 
 
 def test_resolve_namespace_str_truncated(my_future: AbstractFuture):
-    actual = resolve_cache_namespace(
+    actual = determine_cache_namespace(
         "01234567890123456789012345678901234567890123456789extra",
         my_future,
     )
@@ -94,7 +94,7 @@ def test_resolve_namespace_callable_truncated(my_future: AbstractFuture):
     def my_custom_namespace(_: AbstractFuture) -> str:
         return "01234567890123456789012345678901234567890123456789extra"
 
-    actual = resolve_cache_namespace(my_custom_namespace, my_future)
+    actual = determine_cache_namespace(my_custom_namespace, my_future)
     assert actual == "01234567890123456789012345678901234567890123456789"
 
 
@@ -112,27 +112,27 @@ def test_custom_resolve_namespace(
 
         return "whatever"
 
-    actual = resolve_cache_namespace(my_custom_namespace, my_future)
+    actual = determine_cache_namespace(my_custom_namespace, my_future)
     assert actual == "my_namespace"
 
-    actual = resolve_cache_namespace(my_custom_namespace, my_other_future)
+    actual = determine_cache_namespace(my_custom_namespace, my_other_future)
     assert actual == "my_other_namespace"
 
 
 def test_invalid_args_resolve_namespace(my_future: AbstractFuture):
     with pytest.raises(ValueError, match="cannot be None"):
-        resolve_cache_namespace(None, my_future)
+        determine_cache_namespace(None, my_future)
 
-    actual = resolve_cache_namespace("my_namespace", None)  # type: ignore
+    actual = determine_cache_namespace("my_namespace", None)  # type: ignore
     assert actual == "my_namespace"
 
     with pytest.raises(ValueError, match="cannot be None"):
-        resolve_cache_namespace(my_namespace, None)  # type: ignore
+        determine_cache_namespace(my_namespace, None)  # type: ignore
 
     nested_future = mock.MagicMock()
     nested_future.is_root_future.return_value = False
     with pytest.raises(ValueError, match="must be a Resolution root Future"):
-        resolve_cache_namespace(my_namespace, nested_future)
+        determine_cache_namespace(my_namespace, nested_future)
 
 
 def test_malformed_resolve_namespace(my_future: AbstractFuture):
@@ -140,7 +140,7 @@ def test_malformed_resolve_namespace(my_future: AbstractFuture):
         return "my_namespace"
 
     with pytest.raises(TypeError, match="takes 0 positional arguments but 1 was given"):
-        resolve_cache_namespace(my_malformed_namespace, my_future)  # type: ignore
+        determine_cache_namespace(my_malformed_namespace, my_future)  # type: ignore
 
 
 def test_resolve_namespace_error_raised(my_future: AbstractFuture):
@@ -148,4 +148,4 @@ def test_resolve_namespace_error_raised(my_future: AbstractFuture):
         raise ValueError("test error")
 
     with pytest.raises(ValueError, match="test error"):
-        resolve_cache_namespace(my_error_namespace, my_future)
+        determine_cache_namespace(my_error_namespace, my_future)
