+++ sematic/db/models/resolution.py
@@ -1,6 +1,10 @@
 """
 Module defining the Resolution data model.
 
+Note that this class is sometimes referred to as a "Pipeline Run."
+It will be formally renamed to that with
+https://github.com/sematic-ai/sematic/issues/959
+
 Notes regarding container images
 --------------------------------
 In the case of cloud resolution (using `CloudResolver`), the default behavior
@@ -321,3 +325,25 @@ class Resolution(HasUserMixin, Base, JSONEncodableMixin):
         # for the same reason, we can't use value_to_json_encodable, because it imposes
         # the values/types/root_type semantics
         self.git_info_json = json.dumps(dataclasses.asdict(value), sort_keys=True)
+
+
+# Aliases to aid in the rename of resolution -> pipeline run
+# The current code takes the following convention with regards
+# to which variant is used in what places:
+# - Runner code uses the "pipeline run" variants
+# - API client python method names use the "pipeline run" variants
+# - API client HTTP paths use "resolution" API paths
+# - Server-side API supports only "resolution" HTTP paths
+# - DB queries module uses "resolution" name
+# - DB tables use "resolution" name
+#
+# The plan is to progress down this list in order
+# as changes are implemented. For the server-side API
+# endpoints, there will be a transition period where
+# both /api/v1/resolution and api/v1/pipeline_run
+# variants will be accepted, to support backwards
+# compatibility.
+PipelineRunStatus = ResolutionStatus
+InvalidPipelineRun = InvalidResolution
+PipelineRunKind = ResolutionKind
+PipelineRun = Resolution
