+++ sematic/resolvers/silent_resolver.py
@@ -2,16 +2,18 @@
 import logging
 
 # Sematic
+from sematic.resolver import Resolver
 from sematic.runners.silent_runner import SilentRunner
 
 logger = logging.getLogger(__name__)
 
 
-class SilentResolver(SilentRunner):
-    # Stub to allow people time to transition to Runner instead
+class SilentResolver(SilentRunner, Resolver):
+    """A stub to help facilitate the transition from Resolver -> Runner"""
+
     def __init__(self, *args, **kwargs):
         logger.warning(
             "SilentResolver will soon be deprecated in favor of SilentRunner. "
             "Please migrate as soon as you are able."
         )
-        return super().__init__(*args, **kwargs)
+        super().__init__(*args, **kwargs)
